hiscox_integration/api/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
requests==2.26.0
psycopg2-binary==2.9.6
